rename scheduling algorithm
