+++ scheduler/calculate/schedulers/bestCpuMemFit/calculate_test.go
@@ -1,4 +1,4 @@
-package rootBestCpuMemFit
+package bestCpuMemFit
 
 import (
 	"encoding/json"
