+++ scheduler/calculate/schedulers/rootRandom/calculate.go
@@ -1,4 +1,4 @@
-package rootBestCpuMemFit
+package bestCpuMemFit
 
 import (
 	"encoding/json"
